Prodia: types to router
