+++ src/modules/prodia/prodia.router.ts
@@ -2,8 +2,6 @@ import { z } from 'zod';
 
 import { createTRPCRouter, publicProcedure } from '~/modules/trpc/trpc.server';
 
-import { Prodia } from './prodia.types';
-
 
 const imagineInputSchema = z.object({
   prodiaKey: z.string().optional(),
@@ -38,7 +36,7 @@ export const prodiaRouter = createTRPCRouter({
       const tStart = Date.now();
 
       // crate the job, getting back a job ID
-      const jobRequest: Prodia.Wire.Imagine.JobRequest = {
+      const jobRequest: JobRequest = {
         model: input.prodiaModel,
         prompt: input.prompt,
         ...(!!input.cfgScale && { cfg_scale: input.cfgScale }),
@@ -46,7 +44,7 @@ export const prodiaRouter = createTRPCRouter({
         ...(!!input.negativePrompt && { negative_prompt: input.negativePrompt }),
         ...(!!input.seed && { seed: input.seed }),
       };
-      let j: Prodia.Wire.Imagine.JobResponse = await createGenerationJob(prodiaKey, jobRequest);
+      let j: JobResponse = await createGenerationJob(prodiaKey, jobRequest);
 
       // poll the job status until it's done
       let sleepDelay = 2000;
@@ -82,17 +80,40 @@ export const prodiaRouter = createTRPCRouter({
 });
 
 
-function prodiaHeaders(apiKey: string): Record<string, string> {
-  return {
-    'X-Prodia-Key': (apiKey || process.env.PRODIA_API_KEY || '').trim(),
-  };
+export interface JobRequest {
+  model: 'sdv1_4.ckpt [7460a6fa]' | string;
+  prompt: string;
+  // optional, and not even documented, but inferred from the response data
+  cfg_scale?: number;
+  steps?: number;
+  negative_prompt?: string;
+  seed?: number;
 }
 
-async function createGenerationJob(apiKey: string, jobRequest: Prodia.Wire.Imagine.JobRequest): Promise<Prodia.Wire.Imagine.JobResponse> {
+export interface JobResponse {
+  job: string;
+  params: {
+    prompt: string;
+    cfg_scale: number;
+    steps: number;
+    negative_prompt: string;
+    seed: number;
+    upscale: boolean;
+    sampler_name: 'Euler' | string;
+    width: 512 | number;
+    height: 512 | number;
+    options: { sd_model_checkpoint: 'sdv1_4.ckpt [7460a6fa]' | string; };
+  };
+  status: 'queued' | 'generating' | 'succeeded' | 'failed';
+  imageUrl?: string;
+}
+
+
+async function createGenerationJob(apiKey: string, jobRequest: JobRequest): Promise<JobResponse> {
   const response = await fetch('https://api.prodia.com/v1/job', {
     method: 'POST',
     headers: {
-      ...prodiaHeaders(apiKey),
+      'X-Prodia-Key': apiKey,
       'Content-Type': 'application/json',
     },
     body: JSON.stringify(jobRequest),
@@ -104,9 +125,11 @@ async function createGenerationJob(apiKey: string, jobRequest: Prodia.Wire.Imagi
   return await response.json();
 }
 
-async function getJobStatus(apiKey: string, jobId: string): Promise<Prodia.Wire.Imagine.JobResponse> {
+async function getJobStatus(apiKey: string, jobId: string): Promise<JobResponse> {
   const response = await fetch(`https://api.prodia.com/v1/job/${jobId}`, {
-    headers: prodiaHeaders(apiKey),
+    headers: {
+      'X-Prodia-Key': apiKey,
+    },
   });
   if (response.status !== 200) {
     const errorMessage = await response.text() || response.statusText || '' + response.status || 'Unknown Error';
